add: 進行分を更新

+++ module/BlockMotion/InCrossLeft.cpp
@@ -11,7 +11,7 @@ InCrossLeft::InCrossLeft() : BlockMotion(1.23, 1.09) {}
 void InCrossLeft::runLeft(void)
 {
   double targetDistance = 25.0;
-  int targetSpeed = 30;  // 要調整
+  double targetSpeed = 30;  // 要調整
   int angle = 74;
   bool isClockwise = false;  // ピボットターンの方向は左回り
   bool isLeftEdge = true;    // ピボットターン前のエッジは左
